Seperated edge and PS related code
