+++ src/runtime_src/core/edge/drm/zocl/include/zocl_xclbin.h
@@ -45,4 +45,22 @@ int zocl_xclbin_set_dtbo_path(struct drm_zocl_dev *zdev,
 		      struct drm_zocl_slot *slot, char *dtbo_path, uint32_t len);
 int zocl_reset(struct drm_zocl_dev *zdev, const char *buf, size_t count);
 
+int zocl_fpga_mgr_load(struct drm_zocl_dev *zdev, const char *data, int size,
+		   u32 flags);
+int zocl_offsetof_sect(enum axlf_section_kind kind, void *sect,
+	           struct axlf *axlf_full, char __user *xclbin_ptr);
+int zocl_read_sect(enum axlf_section_kind kind, void *sect,
+		   struct axlf *axlf_full, char __user *xclbin_ptr);
+int zocl_update_apertures(struct drm_zocl_dev *zdev, struct drm_zocl_slot *slot);
+void zocl_destroy_cu_slot(struct drm_zocl_dev *zdev, u32 slot_idx);
+int zocl_create_cu(struct drm_zocl_dev *zdev, struct drm_zocl_slot *slot);
+inline bool zocl_xclbin_same_uuid(struct drm_zocl_slot *slot, xuid_t *uuid);
+int zocl_load_sect(struct drm_zocl_dev *zdev, struct axlf *axlf, char __user *xclbin,
+		   enum axlf_section_kind kind, struct drm_zocl_slot *slot);
+int populate_slot_specific_sec(struct drm_zocl_dev *zdev, struct axlf *axlf,
+		   char __user *xclbin, struct drm_zocl_slot *slot);
+bool zocl_bitstream_is_locked(struct drm_zocl_dev *zdev,
+			      struct drm_zocl_slot *slot);
+int zocl_load_partial(struct drm_zocl_dev *zdev, const char *buffer, int length,
+		      struct drm_zocl_slot *slot);
 #endif /* _ZOCL_XCLBIN_H_ */
